Testando se todos os campos estão nos formularios.
close #34
